sketch-agent/requirements.txt
# Core Dependencies
pydantic==2.5.0
pillow==10.1.0
python-dotenv==1.0.0

# Vision AI Providers
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Note: DeepSeek uses OpenAI SDK with base_url override
# Note: Qwen uses OpenAI SDK with base_url override to DashScope

# Optional: Database (if needed for caching)
psycopg2-binary==2.9.9
pgvector==0.2.4

# Optional: API framework (for future expansion)
fastapi==0.109.0
uvicorn==0.27.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
